+++ source/Public/Get-SqlDscTraceFlag.ps1
@@ -45,7 +45,6 @@
 #>
 function Get-SqlDscTraceFlag
 {
-    [System.Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSUseOutputTypeCorrectly', '', Justification = 'Because the rule does not understands that the command returns [System.UInt32[]] when using , (comma) in the return statement')]
     [System.Diagnostics.CodeAnalysis.SuppressMessageAttribute('UseSyntacticallyCorrectExamples', '', Justification = 'Because the rule does not yet support parsing the code when a parameter type is not available. The ScriptAnalyzer rule UseSyntacticallyCorrectExamples will always error in the editor due to https://github.com/indented-automation/Indented.ScriptAnalyzerRules/issues/8.')]
     [OutputType([System.UInt32[]])]
     [CmdletBinding(DefaultParameterSetName = 'ByServerName')]
@@ -83,5 +82,5 @@ function Get-SqlDscTraceFlag
         $script:localizedData.TraceFlag_Get_DebugReturningTraceFlags -f $MyInvocation.MyCommand, ($traceFlags -join ', ')
     )
 
-    return , [System.UInt32[]] $traceFlags
+    return [System.UInt32[]] $traceFlags
 }